requirements.txt
streamlit==1.12.2
gspread==5.5.0
df2gspread==1.0.4
pandas==1.4.2
numpy==1.21.5
plotly==5.6.0
plotly-express==0.4.1 
gspread-dataframe==3.3.0
google==3.0.0
google-api-core==1.25.1
google-api-python-client==1.6.7
google-auth==1.33.0
google-auth-oauthlib==0.5.2
google-cloud-core==1.7.1
google-cloud-storage==1.31.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
PyDrive==1.3.1
Pillow==9.0.1
xlrd==2.0.1
